Change Qauntity to decimal

+++ QTech.App/ReportModels/DriverDeliveryDetail.cs
@@ -17,6 +17,6 @@ namespace QTech.ReportModels
         public string Product { get; set; }
         public string Category { get; set; }
         public decimal ImportPrice { get; set; }
-        public int Qauntity { get; set; }
+        public decimal Qauntity { get; set; }
     }
 }
